requirements.txt
# === LangGraph & LangChain ===
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# === Web Framework ===
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# === Data Validation ===
pydantic>=2.9.0

# === Environment ===
python-dotenv>=1.0.0

# === CORS Support (FastAPI 已包含，但如需额外配置) ===
python-multipart>=0.0.12

# === 数据库与持久化 ===
supabase
psycopg[binary,pool]       # LangGraph 官方推荐的 Postgres 驱动
langchain-postgres         # 方便集成 Postgres 聊天记录